modified:   bt vong lap/7.cpp

+++ bt vong lap/7.cpp
@@ -2,10 +2,10 @@
 #include <iomanip>
 using namespace std;
 
-int f(int n)
+uint64_t f(int n)
 {
     if(n==1) return 1;
-    int s=0;
+    uint64_t s=0;
     for(int i=1;i<n;i++) s+=f(i);
     return n*s;
 }
